refactor(cda-main): move vehicle construction into a private module
Database loading, transport construction, the gateway and the UDS manager were
public items in the crate root, sitting beside the entry points an embedding
application is meant to use. Nothing distinguished the two, so an integration
had no way to tell which items it could rely on.

They cannot be relied on. Their signatures name concrete transport types -
`DoipDiagGateway`, `CanDiagGateway`, `DiagnosticTransportRouter`, `UdsManager` -
so every change to the transport stack changes them, and an integration built on
them breaks on each such refactor. That is not a stability promise anyone can
keep, and it is the specific failure this series exists to stop.

Moving them into a private module makes the boundary structural rather than a
matter of documentation: what remains public in the crate root is what an
embedder may use.
