+++ cda-main/src/cda_factory.rs
@@ -25,7 +25,7 @@ use cda_interfaces::{
 use cda_plugin_security::SecurityPlugin;
 use cda_transport_router::DiagnosticTransportRouter;
 
-use crate::{UdsManagerType, config::configfile::Configuration};
+use crate::{config::configfile::Configuration, vehicle::UdsManagerType};
 
 /// Concrete [`VehicleComponentFactory`] that delegates to
 /// [`crate::create_vehicle_components`].
@@ -82,7 +82,7 @@ where
         >,
         ReloadError,
     > {
-        let crate_components = crate::create_vehicle_components::<SP>(
+        let crate_components = crate::vehicle::create_vehicle_components::<SP>(
             config,
             mdd_paths,
             self.health_providers.as_ref(),
